requirements.txt
matplotlib==3.10.8
greenlet==3.3.2
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
